+++ bsp/stm32/stm32h750-weact-ministm32h7xx/Kconfig
@@ -8,14 +8,14 @@ config BSP_DIR
 config RTT_DIR
     string
     option env="RTT_ROOT"
-    default "rt-thread"
+    default "../../.."
 
 config PKGS_DIR
     string
     option env="PKGS_ROOT"
     default "packages"
- 
+
 source "$RTT_DIR/Kconfig"
 source "$PKGS_DIR/Kconfig"
-source "libraries/Kconfig"
+source "../libraries/Kconfig"
 source "board/Kconfig"
