+++ bsp/stm32/stm32h750-weact-ministm32h7xx/applications/SConscript
@@ -1,12 +1,16 @@
 import rtconfig
 from building import *
+import os
 
-cwd     = GetCurrentDir()
-CPPPATH = [cwd, str(Dir('#'))]
-src     = Split("""
-main.c
-""")
+cwd  = GetCurrentDir()
+path = [cwd]
+src  = Glob('*.c')
 
-group = DefineGroup('Applications', src, depend = [''])
+group = DefineGroup('Applications', src, depend = [''], CPPPATH = path)
+
+list = os.listdir(cwd)
+for item in list:
+    if os.path.isfile(os.path.join(cwd, item, 'SConscript')):
+        group = group + SConscript(os.path.join(item, 'SConscript'))
 
 Return('group')
