+++ bsp/stm32/stm32h750-weact-ministm32h7xx/board/port/drv_spi_flash.c
@@ -7,17 +7,22 @@
  * Date           Author       Notes
  * 2020-08-07     NU-LL        first version
  */
+#include <rtthread.h>
+#include <rtdevice.h>
+#include <drv_spi.h>
+#include <spi_flash.h>
+#include <spi_flash_sfud.h>
 #include <fal.h>
 #include <dfs_fs.h>
 
-// #define DRV_DEBUG
-#define LOG_TAG             "drv.spiflash"
+#define DRV_DEBUG
+#define LOG_TAG                         "drv.spiflash"
 #include <drv_log.h>
 
-#define FS_PARTITION_NAME              "filesystem"
+#define FS_PARTITION_NAME               "filesystem"
 
-#define SPI_CS_GPIO		GPIOD
-#define SPI_CS_PIN		GPIO_PIN_6
+#define SPI_CS_GPIO                     GPIOD
+#define SPI_CS_PIN                      GPIO_PIN_6
 
 static int rt_hw_spi_flash_with_sfud_init(void)
 {
@@ -25,8 +30,9 @@ static int rt_hw_spi_flash_with_sfud_init(void)
 	rt_hw_spi_device_attach("spi1", "spi10", SPI_CS_GPIO, SPI_CS_PIN);
 
     /* init W25Q16    , And register as a block device */
-    if (RT_NULL == rt_sfud_flash_probe("W25Q64", "spi10"))
+    if (RT_NULL == rt_sfud_flash_probe(FAL_USING_NOR_FLASH_DEV_NAME, "spi10"))
     {
+        LOG_E("Failed to probe flash device "FAL_USING_NOR_FLASH_DEV_NAME);
         return -RT_ERROR;
     }
     return err;
@@ -59,7 +65,7 @@ static int mnt(void)
             if (dfs_mount(FS_PARTITION_NAME, "/", "lfs", 0, 0) == 0)
             {
                 mkdir("/qspi",0x777);
-                LOG_I("Filesystem initialized!");
+                LOG_I("mkdir /qspi. Filesystem initialized!");
             }
             else
             {
